+++ src/table.c
@@ -218,7 +218,7 @@ void prefetch_hash_entry(uint64_t hash_key) {
 }
 
 
-void writeHashEntry(uint64_t key, int16_t score, int bestMove, uint8_t depth, uint8_t hashFlag, bool ttPv, board* position) {
+void writeHashEntry(uint64_t key, int16_t score, uint16_t bestMove, uint8_t depth, uint8_t hashFlag, bool ttPv, board* position) {
     // create a TT instance pointer to particular hash entry storing
     // the scoring data for the current board position if available
     tt *hashEntry = &hashTable[get_hash_index(position->hashKey)];
@@ -248,7 +248,7 @@ void writeHashEntry(uint64_t key, int16_t score, int bestMove, uint8_t depth, ui
 }
 
 // read hash entry data
-int readHashEntry(board *position, int *move, int16_t *tt_score,
+int readHashEntry(board *position, uint16_t *move, int16_t *tt_score,
                     uint8_t *tt_depth, uint8_t *tt_flag, bool *tt_pv) {
     // create a TT instance pointer to particular hash entry storing
     // the scoring data for the current board position if available
